requirements.txt
streamlit
langchain
langchain-openai
python-dotenv
psycopg2-binary
openai
redis
pypdf
